+++ src/lib.rs
@@ -57,9 +57,6 @@ mod serialisation;
 pub mod transcript;
 mod util;
 
-#[macro_use]
-extern crate failure;
-
 #[cfg(feature = "nightly")]
 #[doc(include = "../docs/notes-intro.md")]
 pub mod notes {